@@ -11,10 +11,10 @@
   "devDependencies": {
     "@skeletonlabs/skeleton": "^3.1.3",
     "@skeletonlabs/skeleton-svelte": "^1.2.3",
-    "@sveltejs/vite-plugin-svelte": "^6.0.0-next.1",
+    "@sveltejs/vite-plugin-svelte": "^6.0.0-next.2",
     "@tailwindcss/forms": "^0.5.10",
     "@tailwindcss/typography": "^0.5.16",
-    "svelte": "^5.34.7",
+    "svelte": "^5.35.0",
     "vite": "^7.0.0"
   },
   "dependencies": {
